refactor: hide io implementation details
