+++ expertsystem/io/__init__.py
@@ -10,16 +10,17 @@ from pathlib import Path
 
 from expertsystem.amplitude.model import AmplitudeModel
 from expertsystem.particle import ParticleCollection
+from expertsystem.reaction.topology import StateTransitionGraph, Topology
 
-from . import _pdg, dot, xml, yaml
+from . import _dot, _pdg, _xml, _yaml
 
 
 def load_amplitude_model(filename: str) -> AmplitudeModel:
     file_extension = _get_file_extension(filename)
     if file_extension in ["yaml", "yml"]:
-        return yaml.load_amplitude_model(filename)
+        return _yaml.load_amplitude_model(filename)
     if file_extension == "xml":
-        return xml.load_amplitude_model(filename)
+        return _xml.load_amplitude_model(filename)
     raise NotImplementedError(
         f'No parser parser defined for file type "{file_extension}"'
     )
@@ -28,9 +29,9 @@ def load_amplitude_model(filename: str) -> AmplitudeModel:
 def load_particle_collection(filename: str) -> ParticleCollection:
     file_extension = _get_file_extension(filename)
     if file_extension in ["yaml", "yml"]:
-        return yaml.load_particle_collection(filename)
+        return _yaml.load_particle_collection(filename)
     if file_extension == "xml":
-        return xml.load_particle_collection(filename)
+        return _xml.load_particle_collection(filename)
     raise NotImplementedError(
         f'No parser parser defined for file type "{file_extension}"'
     )
@@ -48,13 +49,31 @@ def load_pdg() -> ParticleCollection:
 def write(instance: object, filename: str) -> None:
     file_extension = _get_file_extension(filename)
     if file_extension in ["yaml", "yml"]:
-        return yaml.write(instance, filename)
+        return _yaml.write(instance, filename)
     if file_extension == "xml":
-        return xml.write(instance, filename)
+        return _xml.write(instance, filename)
     if file_extension == "gv":
-        return dot.write(instance, filename)
+        output_str = convert_to_dot(instance)
+        with open(filename, "w") as stream:
+            stream.write(output_str)
+        return None
     raise NotImplementedError(
         f'No writer defined for file type "{file_extension}"'
+    )
+
+
+def convert_to_dot(instance: object) -> str:
+    """Convert a `object` to a DOT language `str`.
+
+    Only works for objects that can be represented as a graph, particularly a
+    `.StateTransitionGraph` or a `list` of `.StateTransitionGraph` instances.
+    """
+    if isinstance(instance, (StateTransitionGraph, Topology)):
+        return _dot.graph_to_dot(instance)
+    if isinstance(instance, list):
+        return _dot.graph_list_to_dot(instance)
+    raise NotImplementedError(
+        f"Cannot convert a {instance.__class__.__name__} to DOT language"
     )
