+++ expertsystem/io/_dot.py
@@ -5,29 +5,7 @@ See :doc:`/usage/visualization` for more info.
 
 from typing import Any, Callable, List, Optional
 
-from expertsystem.reaction.topology import StateTransitionGraph, Topology
-
-
-def convert_to_dot(instance: object) -> str:
-    """Convert a `object` to a DOT language `str`.
-
-    Only works for objects that can be represented as a graph, particularly a
-    `.StateTransitionGraph` or a `list` of `.StateTransitionGraph` instances.
-    """
-    if isinstance(instance, (StateTransitionGraph, Topology)):
-        return __graph_to_dot(instance)
-    if isinstance(instance, list):
-        return __graph_list_to_dot(instance)
-    raise NotImplementedError(
-        f"Cannot convert a {instance.__class__.__name__} to DOT language"
-    )
-
-
-def write(instance: object, filename: str) -> None:
-    output_str = convert_to_dot(instance)
-    with open(filename, "w") as stream:
-        stream.write(output_str)
-
+from expertsystem.reaction.topology import StateTransitionGraph
 
 _DOT_HEAD = """digraph {
     rankdir=LR;
@@ -54,7 +32,7 @@ def embed_dot(func: Callable[[Any], str]) -> Callable[[Any], str]:
 
 
 @embed_dot
-def __graph_list_to_dot(graphs: List[StateTransitionGraph]) -> str:
+def graph_list_to_dot(graphs: List[StateTransitionGraph]) -> str:
     dot_source = ""
     for i, graph in enumerate(graphs):
         dot_source += __graph_to_dot_content(graph, prefix=f"g{i}_")
@@ -62,7 +40,7 @@ def __graph_list_to_dot(graphs: List[StateTransitionGraph]) -> str:
 
 
 @embed_dot
-def __graph_to_dot(graph: StateTransitionGraph) -> str:
+def graph_to_dot(graph: StateTransitionGraph) -> str:
     return __graph_to_dot_content(graph)
