+++ tests/unit/io/test_xml.py
@@ -1,4 +1,4 @@
-"""These tests can be removed when the internal `dict` have been removed."""
+# pylint: disable=protected-access
 
 from expertsystem import io
 from expertsystem.particle import Particle, Spin
@@ -14,7 +14,7 @@ def test_particle():
         isospin=Spin(1.0, -1.0),
         charge=-1,
     )
-    converted_dict = io.xml.object_to_dict(state)
+    converted_dict = io._xml.object_to_dict(state)
     quantum_numbers = converted_dict["QuantumNumber"]
     spin_dict = quantum_numbers[0]
     charge_dict = quantum_numbers[1]
